move business logic to /lib
