+++ word-api/routes/api/words.js
@@ -1,36 +1,9 @@
 'use strict';
 
-const htmlToText = require('html-to-text');
-const requestPage = require('request');
+const Words = require('../../lib/words');
 const validUrl = require('valid-url');
-const wordCount = require('wordcount');
 
 module.exports = function (router) {
-	/**
-	 * Get html from url
-	 *
-	 * @param  {string} url - url of html to scrape.
-	 *
-	 * @return {string} - html.
-	 */
-	function getPage(url) {
-		return new Promise(function (resolve, reject) {
-			let count = 0;
-			requestPage(url, function (error, response, body) {
-				if (error) {
-					return reject(error);
-				} else {
-					let text = htmlToText.fromString(body, {
-						wordwrap: false,
-						ignoreImage: true,
-						ignoreHref: true
-					});
-
-					return resolve(text);
-				}
-			});
-		});
-	}
 
 	/**
 	 * Count words on a webpage
@@ -43,8 +16,8 @@ module.exports = function (router) {
 	function words(request, response) {
 		response.setHeader('Content-Type', 'application/json');
 		if (validUrl.isWebUri(request.query.url)) {
-			getPage(request.query.url).then(function (text) {
-				return response.json({words: wordCount(text)});
+			Words.count(request.query.url).then(function (wordCount) {
+				return response.json({words: wordCount});
 			});
 		} else {
 			return response.status(500).json({error: 'URL was invalid or not supplied'});
